Vehicle-ai/requirements.txt
torch==2.5.0
torchvision==0.20.0
numpy==1.26.4
opencv-python==4.10.0.84
Pillow==10.4.0
pydantic==2.10.4
fastapi==0.115.6
uvicorn==0.34.0
redis==5.2.1
pika==1.3.2
python-multipart==0.0.6
segment-anything==1.0
diffusers==0.21.0
transformers==4.31.0
accelerate==0.21.0
scipy==1.14.1
matplotlib==3.9.3
